fix: harden masterstudy actions per code review
- guard non-array hook response and null course/lesson ids
- share per-action validation (isActionConfigIncomplete) with edit screen
- preserve mapped user email when switching between email actions
- move allActions to common helper (break circular import)
- drop dead handleInput prop from integ layout
