+++ frontend/src/components/AllIntegrations/MasterStudyLms/MasterStudyLms.jsx
@@ -8,43 +8,10 @@ import SnackMsg from '../../Utilities/SnackMsg'
 import Steps from '../../Utilities/Steps'
 import { saveActionConf } from '../IntegrationHelpers/IntegrationHelpers'
 import IntegrationStepThree from '../IntegrationHelpers/IntegrationStepThree'
-import {
-  handleInput,
-  checkMappedFields,
-  isUserEmailMapped,
-  MS_LMS_ACTIONS
-} from './MasterStudyLmsCommonFunc'
+import { isActionConfigIncomplete } from './MasterStudyLmsCommonFunc'
 import MasterStudyLmsAuthorization from './MasterStudyLmsAuthorization'
 import MasterStudyLmsIntegLayout from './MasterStudyLmsIntegLayout'
 import TutorialLink from '../../Utilities/TutorialLink'
-
-export const allActions = [
-  { key: MS_LMS_ACTIONS.COMPLETE_COURSE, label: __('Course complete for the user', 'bit-integrations') },
-  { key: MS_LMS_ACTIONS.COMPLETE_LESSON, label: __('Lesson complete for the user', 'bit-integrations') },
-  { key: MS_LMS_ACTIONS.COMPLETE_QUIZ, label: __('Quiz complete for the user', 'bit-integrations') },
-  { key: MS_LMS_ACTIONS.RESET_COURSE, label: __('Reset user course', 'bit-integrations') },
-  { key: MS_LMS_ACTIONS.RESET_LESSON, label: __('Reset user lesson', 'bit-integrations') },
-  {
-    key: MS_LMS_ACTIONS.ENROLL_USER,
-    label: __('Enroll user in a course', 'bit-integrations'),
-    is_pro: true
-  },
-  {
-    key: MS_LMS_ACTIONS.UNENROLL_USER,
-    label: __('Unenroll user from a course', 'bit-integrations'),
-    is_pro: true
-  },
-  {
-    key: MS_LMS_ACTIONS.MARK_COURSE_COMPLETE,
-    label: __('Mark a course complete for the user', 'bit-integrations'),
-    is_pro: true
-  },
-  {
-    key: MS_LMS_ACTIONS.MARK_LESSON_COMPLETE,
-    label: __('Mark a lesson complete for the user', 'bit-integrations'),
-    is_pro: true
-  }
-]
 
 function MasterStudyLms({ formFields, setFlow, flow, allIntegURL, isInfo, edit }) {
   const navigate = useNavigate()
@@ -73,30 +40,7 @@ function MasterStudyLms({ formFields, setFlow, flow, allIntegURL, isInfo, edit }
   }
 
   function isDisabled() {
-    switch (msLmsConf.mainAction) {
-      case MS_LMS_ACTIONS.COMPLETE_COURSE:
-        return msLmsConf.courseId === undefined
-      case MS_LMS_ACTIONS.RESET_COURSE:
-        return msLmsConf.courseId === undefined
-      case MS_LMS_ACTIONS.COMPLETE_LESSON:
-        return msLmsConf.lessonId === undefined
-      case MS_LMS_ACTIONS.RESET_LESSON:
-        return msLmsConf.lessonId === undefined
-      case MS_LMS_ACTIONS.COMPLETE_QUIZ:
-        return msLmsConf.quizId === undefined
-      case MS_LMS_ACTIONS.ENROLL_USER:
-      case MS_LMS_ACTIONS.UNENROLL_USER:
-      case MS_LMS_ACTIONS.MARK_COURSE_COMPLETE:
-        return msLmsConf.courseId === undefined || !isUserEmailMapped(msLmsConf)
-      case MS_LMS_ACTIONS.MARK_LESSON_COMPLETE:
-        return (
-          msLmsConf.courseId === undefined ||
-          msLmsConf.lessonId === undefined ||
-          !isUserEmailMapped(msLmsConf)
-        )
-      default:
-        return false
-    }
+    return isActionConfigIncomplete(msLmsConf)
   }
 
   return (
@@ -124,7 +68,6 @@ function MasterStudyLms({ formFields, setFlow, flow, allIntegURL, isInfo, edit }
         style={{ ...(step === 2 && { width: 900, height: 'auto', overflow: 'visible' }) }}>
         <MasterStudyLmsIntegLayout
           formFields={formFields}
-          handleInput={e => handleInput(e, msLmsConf, setMsLmsConf, setIsLoading, setSnackbar, formID)}
           msLmsConf={msLmsConf}
           setMsLmsConf={setMsLmsConf}
           isLoading={isLoading}
